+++ packages/examples/intercept_response.ts
@@ -27,7 +27,7 @@
  * @module
  */
 
-import { staticRoute } from "@http/route-deno/static-route";
+import { staticRoute } from "@http/route/static-route";
 import { withFallback } from "@http/route/with-fallback";
 import { interceptResponse } from "@http/interceptor/intercept-response";
 import { skip } from "@http/interceptor/skip";